Renamed component file
